Separate MultipleHives tests into regular FrameworkResolution and MultilevelLookup test classes
